Move Details component to theme-common
