+++ packages/docusaurus-theme-classic/src/theme/Details/styles.module.css
@@ -7,40 +7,5 @@
 
 .details {
   --details-decoration-color: var(--ifm-alert-border-color);
-  --details-summary-arrow-size: 0.5rem;
-}
-
-.details > summary {
-  position: relative;
-  cursor: pointer;
-  list-style: none;
-  margin-left: 1.8rem;
-}
-
-.details > summary::-webkit-details-marker {
-  display: none;
-}
-
-.details > summary:before {
-  position: absolute;
-  content: '';
-  top: 0.25rem;
-  width: 0;
-  height: 0;
-  border-width: var(--details-summary-arrow-size);
-  border-style: solid;
-  border-color: transparent transparent transparent var(--details-decoration-color);
-  transform: translateX(-1.2rem) rotate(0deg);
-  transform-origin: calc(var(--details-summary-arrow-size) / 2) 50%;
-  transition: transform var(--ifm-transition-fast) ease;
-}
-
-.details[data-collapsed='false'] > summary:before {
-  transform: translateX(-1.2rem) rotate(90deg);
-}
-
-.collapsibleContent {
-  margin-top: 1rem;
-  border-top: 1px solid var(--details-decoration-color);
-  padding-top: 1rem;
+  --details-transition: transform var(--ifm-transition-fast) ease;
 }
