+++ packages/docusaurus-theme-common/package.json
@@ -5,8 +5,8 @@
   "main": "./lib/index.js",
   "types": "./lib/index.d.ts",
   "scripts": {
-    "build": "tsc",
-    "watch": "tsc --watch"
+    "build": "node copyUntypedFiles.js && tsc",
+    "watch": "node copyUntypedFiles.js && tsc --watch"
   },
   "publishConfig": {
     "access": "public"
@@ -23,6 +23,8 @@
     "@docusaurus/plugin-content-docs": "2.0.0-beta.3",
     "@docusaurus/plugin-content-pages": "2.0.0-beta.3",
     "@docusaurus/types": "2.0.0-beta.3",
+    "clsx": "^1.1.1",
+    "fs-extra": "^10.0.0",
     "tslib": "^2.1.0"
   },
   "devDependencies": {
